Inline InputMethodMenuController.HardKeyboardListener
This is a follow up CL to our previous CL [1], which introduced

   InputMethodManagerService.HardKeyboardListener,

which was later moved to

   InputMethodMenuController.HardKeyboardListener

as part of refactoring [2].

This CL mechanically inlines it to InputMethodManagerService as it
allows us to achieve the same behavior with fewer lines of code.

This is mechanical refactoring. There should be no observable behavior
change.

 [1]: Ica768083f95c33dc1e494a28ba7d8b6eb989b0ef
      de262ffb3b5e7dff2ae018eff6a2c86824622f2c
 [2]: Ief3c691029b14d848f6ad4a5ec8a8743760c7cda
      e8cd82c3198312bc9393e1365af004f8bce99768

Bug: 192412909
Test: Manually verified by attaching/detaching a hardware keyboard.
Change-Id: I9cce75011a8564c9da8307c5555d87bcba4db2cb