BootReceiver: drop code processing dmesg, employ dmesgd
To remove the need of giving system_server the syslog_read permission,
we are introducing dmesgd, a service that will be launched by init and
perform dmesg output parsing and collection of kernel error reports.

In order to trigger the service, BootReceiver just needs to set the
dmesgd.start property to 1.

Bug: 215095687
Test: run on a user device with injected KFENCE bugs
Change-Id: Ifcfe83ad559633558bba623fb514d7b27d97f6af